+++ packages/web-browser/src/util.ts
@@ -1,22 +1,14 @@
-/**
- * 跳转到新的页面
- * @param url 跳转url
- */
-export function ejectWindow(url: string) {
+export function redirectTo(url: string, target = '_blank') {
   const link = document.createElement('a')
   link.href = url
-  link.target = '_blank'
+  link.target = target
   link.click()
   link.remove()
 }
-/**
- * 拨打手机号
- * @param phoneNumber 手机号码
- */
-export function makePhoneCall(phoneNumber: string) {
-  const link = document.createElement('a')
-  link.style.display = 'none'
-  link.href = `tel:${phoneNumber}`
-  link.click()
-  link.remove()
+
+export function callPhoneNumber(phoneNumber: string) {
+  const aTag = document.createElement('a')
+  aTag.setAttribute('href', `tel:${phoneNumber}`)
+  aTag.setAttribute('target', '_blank')
+  aTag.click()
 }